bot/requirements.txt
python-telegram-bot==22.7
python-dotenv==1.2.2
flask==3.0.0
flask-cors==4.0.0
